+++ packages/catalogue/metadatacatalogue/bundle.js
@@ -26,97 +26,87 @@ Oskari.clazz.define("Oskari.catalogue.bundle.metadatacatalogue.MetadataCatalogue
 			/* service */
 			{
 	            "type": "text/javascript",
-	            "src": "../../../../bundles/catalogue/metadatacatalogue/service/MetadataOptionService.js"
+	            "src": "../../../bundles/catalogue/metadatacatalogue/service/MetadataOptionService.js"
 	        }, {
 	            "type": "text/javascript",
-	            "src": "../../../../bundles/catalogue/metadatacatalogue/service/MetadataSearchService.js"
-	        }, 
-	        /* plugin */
-	        {
-	            "type": "text/javascript",
-	            "src": "../../../../bundles/framework/featuredata2/plugin/MapSelectionPlugin.js"
-	        }, 
-	        /* event */
-	        {
-	            "type": "text/javascript",
-	            "src": "../../../../bundles/catalogue/metadatacatalogue/event/FinishedDrawingEvent.js"
-	        }, 
+	            "src": "../../../bundles/catalogue/metadatacatalogue/service/MetadataSearchService.js"
+	        },
 	        /* instance */
 	        {
 				"type" : "text/javascript",
-				"src" : "../../../../bundles/catalogue/metadatacatalogue/instance.ol2.js"
+				"src" : "../../../bundles/catalogue/metadatacatalogue/instance.js"
 			}, 
 			/* request */
 	        {
 	            "type": "text/javascript",
-	            "src": "../../../../bundles/catalogue/metadatacatalogue/request/AddSearchResultActionRequest.js"
+	            "src": "../../../bundles/catalogue/metadatacatalogue/request/AddSearchResultActionRequest.js"
 	        }, {
 	            "type": "text/javascript",
-	            "src": "../../../../bundles/catalogue/metadatacatalogue/request/AddSearchResultActionRequestHandler.js"
-	        },
+	            "src": "../../../bundles/catalogue/metadatacatalogue/request/AddSearchResultActionRequestHandler.js"
+			},
 	        /* css */
 			{
 				"type" : "text/css",
-				"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/css/style.css"
+				"src" : "../../../bundles/catalogue/metadatacatalogue/resources/css/style.css"
 			}
 		], 
 		"locales" : [  
 		{
 			"lang" : "fi",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/fi.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/fi.js"
 		},{
 			"lang" : "fr",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/fr.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/fr.js"
 		},{
 			"lang" : "en",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/en.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/en.js"
 		},{
 			"lang" : "et",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/et.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/et.js"
 		},{
 			"lang" : "it",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/it.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/it.js"
 		},{
 			"lang" : "is",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/is.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/is.js"
 		},{
 			"lang" : "sl",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/sl.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/sl.js"
 		},{
 			"lang" : "nb",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/nb.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/nb.js"
 		},{
 			"lang" : "nn",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/nn.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/nn.js"
 		},{
 			"lang" : "nl",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/nl.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/nl.js"
 		},{
 			"lang" : "nn",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/nn.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/nn.js"
 		},{
 			"lang" : "sk",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/sk.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/sk.js"
 		},{
 			"lang" : "sv",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/sv.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/sv.js"
 		},{
 			"lang" : "es",
 			"type" : "text/javascript",
-			"src" : "../../../../bundles/catalogue/metadatacatalogue/resources/locale/es.js"
+			"src" : "../../../bundles/catalogue/metadatacatalogue/resources/locale/es.js"
 		}],
 		"resources" : []
 	},