+++ bundles/framework/divmanazer/extension/ExtraFlyout.js
@@ -6,15 +6,17 @@
  *
  */
 Oskari.clazz.define('Oskari.userinterface.extension.ExtraFlyout',
-	/**
+    /**
      * @method create called automatically on construction
      * @static
      *
      * Always extend this class, never use as is.
      */
     function (title, options) {
-    	// UI text for title
+        // UI text for title
         this.title = title;
+        this._visible = false;
+        this._popup = null;
 
         /* @property container the DIV element */
         this.container = null;
@@ -22,10 +24,7 @@ Oskari.clazz.define('Oskari.userinterface.extension.ExtraFlyout',
 
         this.__render();
         Oskari.makeObservable(this);
-    },
-    {
-	    __visible: false,
-		__popup: null,
+    }, {
 	    __templates : {
 	    	popup: jQuery('<div class="oskari-flyout">' +
 	    		'<div class="oskari-flyouttoolbar">' +
@@ -46,130 +45,129 @@ Oskari.clazz.define('Oskari.userinterface.extension.ExtraFlyout',
                 '</div>')
         },
 	    isVisible : function() {
-	    	return this.__visible;
+	    	return this._visible;
 	    },
 	    show: function(){
 	    	var me = this;
-	    	me.__popup.show();
-    		me.__visible = true;
+	    	me._popup.show();
+    		me._visible = true;
     		this.trigger('show');
 	    },
 	    hide: function(suppressEvent){
 	    	var me = this;
-	    	me.__popup.hide();
-    		me.__visible = false;
+	    	me._popup.hide();
+    		me._visible = false;
     		suppressEvent = suppressEvent ? suppressEvent: false;
+            if(!suppressEvent) {
+                this.trigger('hide');
+            }
+        },
+        __render: function(){
+            var me = this;
+            var popup = me._popup || me.__templates.popup.clone();
 
-    		if(!suppressEvent) {
-    			this.trigger('hide');
-    		}
-	    },
-	    __render: function(){
-	    	var me = this;
-	        var popup = me.__popup || me.__templates.popup.clone();
-
-	        if(!me.__popup) {
+            if(!me._popup) {
                 if(!me.options.container) {
-	        	  jQuery('body').append(popup);
+                  jQuery('body').append(popup);
                 } else {
                     me.options.container.append(popup);
                 }
-	        	popup.find('.icon-close').bind('click', function(){
-	        		me.hide();
-	        	});
-	        	me.__popup = popup;
-	        	me.hide(true);
-	    	}
+                popup.find('.icon-close').bind('click', function(){
+                    me.hide();
+                });
+                me._popup = popup;
+                me.hide(true);
+            }
 
-	    	if(me.options.visible === true) {
-	    		me.show();
-	    	}
+            if(me.options.visible === true) {
+                me.show();
+            }
 
-	    	me.setTitle(me.title);
-	    	me.addClass(me.options.cls);
-			me.setSize(me.options.width, me.options.height);
-	    },
-	    setTitle: function(title) {
-	    	var me = this;
-	    	if(!this.__popup) {
-	    		return;
-	    	}
-	    	me.__popup.find('.oskari-flyout-title p').html(title || '');
-	    },
-	    getTitle: function() {
-	    	var me = this;
-	    	return me.__popup.find('.oskari-flyout-title p');
-	    },
-	    /**
-	     * @method  @public setContent Set content
-	     * @param {Object} content jQuery object
-	     */
-	    setContent: function(content) {
-	    	var me = this;
-	    	me.__popup.find('.oskari-flyoutcontentcontainer').html(content);
-	    },
-	    addClass: function(cls) {
-	    	if(!this.__popup) {
-	    		return;
-	    	}
-	    	this.__popup.addClass(cls);
-	    },
-	    setSize : function(width, height) {
-	    	if(!this.__popup) {
-	    		return;
-	    	}
-	    	if(width) {
-	    		this.__popup.css('width', width);
-	    	}
-	    	if(height) {
-	    		this.__popup.css('height', height);
-	    	}
+            me.setTitle(me.title);
+            me.addClass(me.options.cls);
+            me.setSize(me.options.width, me.options.height);
+        },
+        setTitle: function(title) {
+            var me = this;
+            if(!this._popup) {
+                return;
+            }
+            me._popup.find('.oskari-flyout-title p').html(title || '');
+        },
+        getTitle: function() {
+            var me = this;
+            return me._popup.find('.oskari-flyout-title p');
+        },
+        /**
+         * @method  @public setContent Set content
+         * @param {Object} content jQuery object
+         */
+        setContent: function(content) {
+            var me = this;
+            me._popup.find('.oskari-flyoutcontentcontainer').html(content);
+        },
+        addClass: function(cls) {
+            if(!this._popup) {
+                return;
+            }
+            this._popup.addClass(cls);
+        },
+        setSize : function(width, height) {
+            if(!this._popup) {
+                return;
+            }
+            if(width) {
+                this._popup.css('width', width);
+            }
+            if(height) {
+                this._popup.css('height', height);
+            }
 
-	    },
-	    bringToTop : function() {
-	    	if(!this.__popup) {
-	    		return;
-	    	}
-            this.__popup.css('z-index', 20000);
-	    },
-	    move : function(left, top, keepOnScreen) {
-	    	if(!this.__popup) {
-	    		return;
-	    	}
-	    	if(keepOnScreen) {
-	    		var size = this.getSize();
-	            if(left + size.width > jQuery(window).width()) {
-	                left = jQuery(window).width() - size.width;
-	            }
-	            if(left < 0) {
-	                left = 0;
-	            }
-	            if(top + size.height > jQuery(window).height()) {
-	                top = jQuery(window).height() - size.height;
-	            }
-	            if(top < 0) {
-	                top = 0;
-	            }
-	    	}
-            this.__popup.css({
+        },
+        bringToTop : function() {
+            if(!this._popup) {
+                return;
+            }
+            this._popup.css('z-index', 20000);
+        },
+        move : function(left, top, keepOnScreen) {
+            if(!this._popup) {
+                return;
+            }
+            if(keepOnScreen) {
+                var size = this.getSize();
+                if(left + size.width > jQuery(window).width()) {
+                    left = jQuery(window).width() - size.width;
+                }
+                if(left < 0) {
+                    left = 0;
+                }
+                if(top + size.height > jQuery(window).height()) {
+                    top = jQuery(window).height() - size.height;
+                }
+                if(top < 0) {
+                    top = 0;
+                }
+            }
+            this._popup.css({
                 left: left,
                 top: top
             });
-	    },
-	    getPosition : function() {
-	    	if(!this.__popup) {
-	    		return;
-	    	}
-	    	return this.__popup.position();
-	    },
-	    getSize : function() {
-	    	if(!this.__popup) {
-	    		return;
-	    	}
-	    	return {
-	    		width : this.__popup.outerWidth(),
-	    		height : this.__popup.height()
-	    	};
+        },
+        getPosition : function() {
+            if(!this._popup) {
+                return;
+            }
+            return this._popup.position();
+        },
+        getSize : function() {
+            if(!this._popup) {
+                return;
+            }
+            return {
+                width : this._popup.outerWidth(),
+                height : this._popup.height()
+            };
         },
         /**
          * @method makeDraggable
@@ -182,8 +180,8 @@ Oskari.clazz.define('Oskari.userinterface.extension.ExtraFlyout',
                 scroll: false,
                 handle: '.oskari-flyouttoolbar'
             };
-            me.__popup.css('position', 'absolute');
-            me.__popup.draggable(dragOptions);
+            me._popup.css('position', 'absolute');
+            me._popup.draggable(dragOptions);
         },
         getElement: function(){
             return this.__popup;
@@ -200,8 +198,8 @@ Oskari.clazz.define('Oskari.userinterface.extension.ExtraFlyout',
         },
         _calcSideLabelPositions: function(){
             var me = this;
-            var sidelabels = me.__popup.find('.sidetool');
-            var flyout = me.__popup;
+            var sidelabels = me._popup.find('.sidetool');
+            var flyout = me._popup;
             var heights = flyout.find('.oskari-flyouttoolbar').outerHeight();
             jQuery.each(sidelabels.get(), function(index, sidelabel) {
                 if(index === 0) {
@@ -235,7 +233,7 @@ Oskari.clazz.define('Oskari.userinterface.extension.ExtraFlyout',
 
             var textSize = textWidth(sidelabel.find('label'));
 
-            var flyout = me.__popup;
+            var flyout = me._popup;
             var parent = flyout.parent();
 
             sidelabel.css('margin-left', '-16px');
@@ -245,7 +243,7 @@ Oskari.clazz.define('Oskari.userinterface.extension.ExtraFlyout',
 
             if(typeof callback === 'function') {
                 sidelabel.on('click', function() {
-                    var position = me.__popup.position();
+                    var position = me._popup.position();
                     var bounds = {
                         left : position.left + sidelabel.position().left - 16,
                         top : position.top + sidelabel.position().top
@@ -259,7 +257,7 @@ Oskari.clazz.define('Oskari.userinterface.extension.ExtraFlyout',
             me._calcSideLabelPositions();
 
             if(!me._addedResizeListener){
-                me.__popup.bind('DOMSubtreeModified', function(){
+                me._popup.bind('DOMSubtreeModified', function(){
                     clearTimeout(me._sidetoolTimer);
                     me._sidetoolTimer = setTimeout(function(){
                         me._calcSideLabelPositions();
@@ -273,9 +271,10 @@ Oskari.clazz.define('Oskari.userinterface.extension.ExtraFlyout',
          */
         removeSideTools: function(){
             var me = this;
-            var sidelabels = me.__popup.find('.sidetool');
+            var sidelabels = me._popup.find('.sidetool');
             sidelabels.each(function(index, sidelabel) {
                 jQuery(sidelabel).remove();
             });
+
         }
 });