+++ api/CHANGELOG.md
@@ -63,6 +63,10 @@ Notes:
 - Only lines and polygons are counted for the area/length (circles/points with buffers are not).
 - The measurements are for non-buffered features.
 
+### metadatacatalogue [add]
+
+Added new OpenLayers independent version of metadatacatalogue. New bundle.js is under packages/catalogue/metadatacatalogue/. In the new version `MetaData.FinishedDrawingEvent` is removed. Instead you can listen to `DrawingEvent` where id "catalogue.bundle.metadatacatalogue" and isFinished is true.
+
 ## 1.44
 
 ### [mod] [breaking] AddLayerListFilterRequest